Using CSS Modules to split CSS in different components.
This also necessitated the creation of the Button, which is either
formed using <button> or <a> depending on whether it is passed an href
argument.
