+++ src/app/app.module.css
@@ -39,37 +39,6 @@ div.buttonpadding {
     padding-block: 10px;
 }
 
-a.centre,
-button.centre {
-    background-color: transparent;
-    color: rgb(var(--colour-pearl));
-    padding: 10px;
-    min-width: 100px;
-    font-weight: bold;
-    font-size: 20px;
-    display: block;
-    width: 300px;
-    margin-inline: auto;
-    border-radius: 10px;
-    cursor: pointer;
-    border: 2px solid rgb(var(--colour-pearl));
-    box-shadow: 0 3px rgb(var(--colour-pearl));
-    text-align: center;
-    text-decoration: none;
-}
-
-a.centre:hover,
-button.centre:hover {
-    opacity: 0.8;
-}
-
-div.centre {
-    margin-inline: auto;
-    width: 800px;
-    display: block;
-    text-align: center;
-}
-
 svg.padded {
     padding-block: 20px;
 }
